backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-core==2.23.4

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
cryptography==42.0.5

# HTTP Client & Network
httpx==0.27.2
python-multipart==0.0.12

# Environment & Config
python-dotenv==1.0.1

# Database (for future use)
sqlalchemy==2.0.28
psycopg2-binary==2.9.9

# Data Processing
pandas==2.2.1
numpy==1.26.4

# Monitoring & Logging
python-json-logger==2.0.7

# Firebase Admin SDK
firebase-admin==6.4.0
